+++ Frontend/AppSettings.config
@@ -5,13 +5,10 @@
     <add key="Environment" value="local" />
     <add key="ClientValidationEnabled" value="true" />
     <add key="UnobtrusiveJavaScriptEnabled" value="true" />
+    
     <add key="MongoUrl" value="mongodb://mac.localhost" />
-    <add key="EventStoreIp" value="127.0.0.1"/>
-    <add key="EventStoreTcpPort" value="1113"/>
-    <add key="EventStoreUserName" value="admin"/>
-    <add key="EventStorePassword" value="changeit"/>
-    <add key="UseEventStore" value="false"/>
-    <add key="RedisIp" value="logging-pp.otenv.com"/>
+    
+    <add key="RedisHost" value="logging-pp.otenv.com"/>
     <add key="RedisPort" value="6379"/>
-    <add key="RedisMaxRetries" value="3"/>
+    <add key="RedisRetryTime" value="3"/>
 </appSettings>
