+++ Frontend/App_Start/ServiceBusConfig.cs
@@ -1,12 +1,12 @@
-﻿using System;
-using System.Configuration;
-using System.IO;
-using MongoDataAccess;
-using TaskScheduler;
+﻿using MongoDataAccess;
+using TaskScheduler.Configuration;
 using TaskScheduler.EventBus;
 using TaskScheduler.EventHandlers;
 using TaskScheduler.Logging;
 using TaskScheduler.Operations;
+using TaskScheduler.Repositories;
+using TaskScheduler.Services;
+using TaskScheduler.Services.Files;
 
 namespace Frontend
 {
@@ -14,29 +14,23 @@ namespace Frontend
     {
         public static void Initialize()
         {
+            var configuration = new Configuration();
             var eventHandler = new EventHandlerFactory();
-            var fileName = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "TaskSchedulerConfiguration.json");
-            var config = File.OpenText(fileName).ReadToEnd();
-            var mongourl = ConfigurationManager.AppSettings["MongoUrl"];
-            var mongoTaskRepository = new MongoTaskRepository(mongourl);
-
-            var logger = GetLogger();
+            var mongoTaskRepository = new MongoTaskRepository(configuration);
+            var logger = GetLogger(configuration);
 
             eventHandler.RegisterInstance(() => new UpdateTaskResponseStatusEventHandler(mongoTaskRepository));
-            eventHandler.RegisterInstance(() => new RunTaskEventHandler(new StandardDateTimeProvider(), mongoTaskRepository, new TimeSpanEvaluator(), new OperationResolver(), logger));
-            eventHandler.RegisterInstance(() => new ElapsedTimeEventHandler(mongoTaskRepository, new StandardDateTimeProvider()));
-            eventHandler.RegisterInstance(() => new InitializeTaskEventHandler(new TimeSpanEvaluator(), new StandardDateTimeProvider(), mongoTaskRepository));
-            eventHandler.RegisterInstance(() => new InitializeTaskManagerEventHandler(new JSonConfigurationRepository(config), mongoTaskRepository));
-            eventHandler.RegisterInstance(() => new ErrorThrownEventHandler(new MongoErrorLogRepository(mongourl)));
+            eventHandler.RegisterInstance(() => new RunTaskEventHandler(new DateService(), mongoTaskRepository, new OperationResolver(), logger));
+            eventHandler.RegisterInstance(() => new ElapsedTimeEventHandler(mongoTaskRepository, new DateService()));
+            eventHandler.RegisterInstance(() => new InitializeTaskEventHandler(new DateService(), mongoTaskRepository));
+            eventHandler.RegisterInstance(() => new InitializeTaskManagerEventHandler(new ConfigurationRepository(configuration, new FileService()), mongoTaskRepository));
+            eventHandler.RegisterInstance(() => new ErrorThrownEventHandler(new MongoErrorLogRepository(configuration)));
             Bus.InitializeBus(eventHandler, logger);
         }
 
-        private static IRedisLogger GetLogger()
+        private static IRedisLogger GetLogger(IConfiguration configuration)
         {
-            var redisIp = ConfigurationManager.AppSettings["RedisIp"];
-            var redisPort = int.Parse(ConfigurationManager.AppSettings["RedisPort"]);
-            var redisMaxRetries = int.Parse(ConfigurationManager.AppSettings["RedisMaxRetries"]);
-            var logger = new RedisLogger(new RedisConnectionFactory(new RedisConnectionWrapper(), redisIp, redisPort, redisMaxRetries), "logstash");
+            var logger = new RedisLogger(new RedisConnectionFactory(new RedisConnectionWrapper(), configuration.Logger), "logstash");
             return logger;
         }
     }
