+++ MongoDataAccess/MongoErrorLogRepository.cs
@@ -1,19 +1,21 @@
 ﻿using MongoDB.Driver;
 using TaskScheduler;
+using TaskScheduler.Configuration;
+using TaskScheduler.Repositories;
 
 namespace MongoDataAccess
 {
     public class MongoErrorLogRepository : IErrorLogRepository
     {
-        private readonly string _mongoUrl;
-
-        public MongoErrorLogRepository(string mongoUrl)
-        {
-            _mongoUrl = mongoUrl;
-        }
-
         private const string CollectionName = "ErrorLog";
         private const string DbName = "TaskScheduler";
+
+        private readonly string _mongoUrl;
+
+        public MongoErrorLogRepository(IConfiguration configuration)
+        {
+            _mongoUrl = configuration.Mongo.Address;
+        }
 
         public void Save(ErrorEntry errorEntry)
         {
