--- TaskScheduler/IDateTimeProvider.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace TaskScheduler
-{
-    public interface IDateTimeProvider
-    {
-        DateTime NowUtc { get; }
-    }
-}
