--- TaskScheduler/IErrorLogRepository.cs
@@ -1,7 +1,0 @@
-﻿namespace TaskScheduler
-{
-    public interface IErrorLogRepository
-    {
-        void Save(ErrorEntry entry);
-    }
-}
