--- TaskScheduler/ITimeSpanEvaluator.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace TaskScheduler
-{
-    public interface ITimeSpanEvaluator
-    {
-        DateTime Evaluate(DateTime utcNow, string utcRunningTime);
-    }
-}
