+++ TaskScheduler/Repositories/IErrorLogRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace TaskScheduler.Repositories
+{
+    public interface IErrorLogRepository
+    {
+        void Save(ErrorEntry entry);
+    }
+}
