+++ TaskScheduler/Services/Files/FileService.cs
@@ -1,0 +1,12 @@
+﻿using System.IO;
+
+namespace TaskScheduler.Services.Files
+{
+    public class FileService : IFileService
+    {
+        public string GetContents(string filePath)
+        {
+            return File.OpenText(filePath).ReadToEnd();
+        }
+    }
+}
