+++ TaskScheduler/Services/Files/IFileService.cs
@@ -1,0 +1,7 @@
+﻿namespace TaskScheduler.Services.Files
+{
+    public interface IFileService
+    {
+        string GetContents(string filePath);
+    }
+}
