+++ TaskScheduler/Services/IDateService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace TaskScheduler
+{
+    public interface IDateService
+    {
+        DateTime Evaluate(DateTime utcNow, string utcRunningTime);
+
+        DateTime NowUtc { get; }
+    }
+}
