+++ TaskScheduler/Services/TaskMonitoringService.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using TaskScheduler.EventBus;
 using TaskScheduler.Events;
+using TaskScheduler.Repositories;
 
 namespace TaskScheduler.Services
 {
