--- TaskScheduler/StandardDateTimeProvider.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace TaskScheduler
-{
-    public class StandardDateTimeProvider : IDateTimeProvider
-    {
-        public DateTime NowUtc
-        {
-            get
-            {
-                return DateTime.UtcNow;
-            }
-        }
-    }
-}
